move client sheet
